+++ docs/changes.md
@@ -1,5 +1,13 @@
 # Changes
 
+## 0.2.0
+
+* `ErrorInfo` JSON serialization/deserialization.
+* ErrorInfoMiddleware to write collected errors to JSON files.
+* New Err.setup options:
+  * `error_info_path`
+  * `error_info_compress`
+
 ## 0.1.1
 
 * `__version__` attribute.